remove dead code and fix compiler warnings
