+++ src/code_gen/mod.rs
@@ -1,24 +1,10 @@
 use instruction::{Instruction, StackValue};
 
+use crate::code_gen::error::CodeGenError;
 use crate::parser::{Method, Service, Statement};
 
 pub mod error;
 pub mod instruction;
-
-#[derive(Debug, Clone)]
-pub enum CodeGenError {
-    InvalidStatement(String),
-}
-
-impl std::fmt::Display for CodeGenError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            CodeGenError::InvalidStatement(msg) => write!(f, "Invalid statement: {}", msg),
-        }
-    }
-}
-
-impl std::error::Error for CodeGenError {}
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum PrintType {
