+++ src/vm_coordinator.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use opentelemetry::trace::{Span, SpanKind, Status, Tracer};
 use opentelemetry::{global, trace::TracerProvider as _, KeyValue};
-use opentelemetry_sdk::trace::{self, SdkTracerProvider};
+use opentelemetry_sdk::trace::SdkTracerProvider;
 use opentelemetry_semantic_conventions::resource::SERVICE_NAME;
 use tokio::sync::mpsc;
